update: create ffi::compat to store compatibility shims
