+++ src/types/dict.rs
@@ -248,7 +248,9 @@ impl<'py> PyDictMethods<'py> for Bound<'py, PyDict> {
         ) -> PyResult<Option<Bound<'py, PyAny>>> {
             let py = dict.py();
             let mut result: *mut ffi::PyObject = std::ptr::null_mut();
-            match unsafe { ffi::PyDict_GetItemRef(dict.as_ptr(), key.as_ptr(), &mut result) } {
+            match unsafe {
+                ffi::compat::PyDict_GetItemRef(dict.as_ptr(), key.as_ptr(), &mut result)
+            } {
                 std::os::raw::c_int::MIN..=-1 => Err(PyErr::fetch(py)),
                 0 => Ok(None),
                 1..=std::os::raw::c_int::MAX => Ok(Some(unsafe { result.assume_owned(py) })),
